+++ backend/game_data.h
@@ -5,6 +5,6 @@ namespace GameData
 { // functiion for getter game's data
         PowerPlantStats GetPowerPlantStats(PlantType type, int level);
         ResourcePlantStats GetResourcePlantStats(ResourceType type, int level);
-        CityData CreateCityData(CityType type, float min, float max, float charge);
-        float GetTranportRate(TransportType item);
+        CityData CreateCityData(CityType type, double min, double max, double charge);
+        double GetTranportRate(TransportType item);
 };
